Move some lexer functions to misc module
